pyproject.toml
[tool.poetry]
name = "traffic-rules-mvp"
version = "0.1.0"
description = "Red light stop anomaly detection MVP skeleton"
authors = ["AI Product Team <team@example.com>"]
license = "Proprietary"
readme = "README.md"
packages = [{ include = "traffic_rules", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
torch = "2.4.1"
torchvision = "0.19.1"
torchaudio = "2.4.1"
pydantic = "^2.7.0"
pyyaml = "^6.0.0"
typer = { version = "^0.12.0", extras = ["all"] }
rich = "^13.7.0"
prometheus-client = "^0.20.0"
opencv-python = "^4.10.0"
numpy = "^2.1.0"
networkx = "^3.2.0"
matplotlib = "^3.9.0"
tqdm = "^4.66.0"
structlog = "^24.1.0"
pandas = "^2.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-cov = "^5.0.0"
black = { version = "24.8.0", allow-prereleases = true }
isort = "^5.13.0"
ruff = "^0.6.0"
mypy = "^1.11.0"
bandit = "^1.7.0"
pip-audit = "^2.7.0"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "B", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
strict = true
python_version = "3.11"
warn_unused_configs = true

[tool.pytest.ini_options]
addopts = "-ra --maxfail=1"
testpaths = ["tests"]

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "primary"

